Simplify train and predict
