+++ sagify/template/sagify/training/training.py
@@ -1,0 +1,18 @@
+def train(input_data_path, model_save_path, hyperparams_path=None):
+    """
+    The function to execute the training.
+
+    :param input_data_path: [str], input directory path where all the training file(s) reside in
+    :param model_save_path: [str], directory path to save your model(s)
+    :param hyperparams_path: [optional[str], default=None], input path to hyperparams json file.
+    Example:
+        {
+            "max_leaf_nodes": 10,
+            "n_estimators": 200
+        }
+    """
+    # TODO: If exists, read in hyperparams file JSON content
+
+    # TODO: Write your modeling logic
+
+    # TODO: save the model(s) under 'model_save_path'
