+++ tests/commands/test_initialize.py
@@ -29,9 +29,11 @@ def test_init_happy_case():
             assert os.path.isfile('src/sagify/Dockerfile')
             assert os.path.isfile('src/sagify/training/__init__.py')
             assert os.path.isfile('src/sagify/training/train')
+            assert os.path.isfile('src/sagify/training/training.py')
             assert os.path.isfile('src/sagify/prediction/__init__.py')
             assert os.path.isfile('src/sagify/prediction/nginx.conf')
             assert os.path.isfile('src/sagify/prediction/predictor.py')
+            assert os.path.isfile('src/sagify/prediction/prediction.py')
             assert os.path.isfile('src/sagify/prediction/wsgi.py')
             assert os.path.isfile('src/sagify/prediction/serve')
             assert os.path.isfile('src/sagify/local_test/train_local.sh')
